Added space for custom implementations - new CustomAgents and CustomAlgorithms classes are used to configure and instantiate custom agents - new Algorithm class now holds the implementations
