+++ src/rl_framework/agent/__init__.py
@@ -1,3 +1,3 @@
 from .agent import Agent
+from .custom_agent import CustomAgent, CustomAlgorithm
 from .stable_baselines import StableBaselinesAgent, StableBaselinesAlgorithm
-from .q_learning import QLearningAgent
